refactor(common)!: move BaseHoodieTableFileIndex to org.apache.hudi.core.read
Relocates the engine-agnostic file index base (and its unit test) from
the org.apache.hudi root package to org.apache.hudi.core.read,
establishing the read-side home in the core namespace for domain-aware
query/listing abstractions (Phase 4 read-path work lands alongside).
It is a consumer of the metadata subsystem, not part of it, so
org.apache.hudi.metadata was deliberately not chosen.

BREAKING CHANGE: source/binary break for external code importing
BaseHoodieTableFileIndex or its nested types (PartitionPath,
FileStatusCache) directly. Engine subclasses shipped by Hudi keep
their FQNs (HiveHoodieTableFileIndex, SparkHoodieTableFileIndex), and
engine bundles are self-consistent, so existing Presto/Trino
deployments are unaffected at runtime; connector source trees that
import the base class directly see a compile-time import fix at their
next Hudi version bump. Never a config value or reflection target.

The move surfaced several same-package couplings, now explicit: the
spark-common scala classes in package org.apache.hudi (file index and
index-support family) referenced the base bare and now import it,
HoodieBaseRelation read the protected metadataConfig field directly
and now uses the getter, and the read-only introspection methods
getAllQueryPartitionPaths/areAllPartitionPathsCached used by tests
across the old package boundary are now public.

Verified across common, hadoop-common, hadoop-mr, spark modules and
utilities with -Dspark4.1; file index suites pass.
